+++ docs/docs/changelog.md
@@ -5,10 +5,15 @@ group:
 
 # 更新内容
 
+# Todo
+
+- [ ] 使用useIntersection 优化虚拟滚动的hook
+
 ## 待更新
 
 #### Feature
 
+- 增加useMutationObserver
 - 文档重新梳理hook分类
 - useInterval改为useIntervalFn