+++ docs/docs/changelog.md
@@ -20,7 +20,7 @@ group:
 #### Fixed
 
 - 修复debounce、throttle引用问题
-- useIsScroll，如果内部dom变化，不会正确判断是否处于滚动装填
+- useIsScroll，如果内部dom变化，不会正确判断是否处于滚动状态
 
 
 # 更新日志